Move login logic into useAuth hook and access with unstated-next
