+++ src/hooks/auth.tsx
@@ -54,6 +54,45 @@ export const useAuth = () => {
 
   //====== signup ends ==========
 
+
+  //====== Login hook code ========
+
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const loginUser = async (e: any) => {
+    e.preventDefault();
+    const response = await fetch(`${process.env.REACT_APP_BASE_URL}/auth/login`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        email,
+        password
+      })
+    });
+
+    const data = await response.json();
+
+    // TODO: This works but include a property with boolean value of true/false in the API for use here instead
+    if (data.user) {
+      navigate('/books');
+    }
+
+    console.log(data);
+  }
+
+  const handleEmail = (eventTargetValue: string) => {
+    setEmail(eventTargetValue);
+  }
+
+  const handlePassword = (eventTargetValue: string) => {
+    setPassword(eventTargetValue);
+  }
+
+  //====== login ends ==========
+
   return {
     signinEmail,
     signinPassword,
@@ -63,7 +102,13 @@ export const useAuth = () => {
     handleSigninEmail,
     handleSigninPassword,
     handlePasswordConfirmation,
-    handleSigninName
+    handleSigninName,
+
+    email,
+    password,
+    loginUser,
+    handleEmail,
+    handlePassword
   }
 }
